refactor: #107 Step 5 reducer 分割 + dead code 削除 + reducer テスト + 動画演出フック集約 (#126)
* refactor: Step 5 — reducer を src/hooks/card-shogi/reducer.ts に分離

use-card-shogi-game.ts は 1,070 行に肥大化しており、純粋な reducer
ロジックと React hook 副作用 (useEffect / useCallback / DB 永続化) が
混在していた。Step 5 (Issue #107) として reducer 関数本体 + 内部 helper
+ Action 型を src/hooks/card-shogi/reducer.ts に切り出し、フック側は
useReducer + useEffect + 公開 API のみを担う薄い shell に縮小。

- src/hooks/card-shogi/reducer.ts (新規 708 行):
  - ShogiAction / Action 型 (export)
  - CardShogiGameStateInternal 型 (export)
  - makeMoveWithEffects / isKingInCheckAfterMove (内部 helper)
  - reducer 関数本体 (export)
- use-card-shogi-game.ts: 1,070 → 311 行 (71% 縮減)
  - reducer に必要な imports (effects.ts / definitions.ts / board.ts /
    moves.ts / rules.ts) を新ファイル側に集約し、フック側は
    moveToNotation / getAiMove / saveCardShogiMove / isValidCardTargetSquare
    のみに削減

ロジックは 1 行も変更なし (move-only リファクタ)。export shape も
useCardShogiGame のみで不変。

Issue #107 Step 5

* refactor: Step 5 — CardAction の dead code を削除 (CHARGE_MANA / SET_TRAP / TRIGGER_TRAP)

旧 CardAction には CHARGE_MANA / SET_TRAP / TRIGGER_TRAP の 3 つが
定義されていたが、grep 確認した結果 UI からも内部からも dispatch
されていない dead code だった (reducer 内でも use case がなかった)。

実際の動作経路:
- マナチャージ: MAKE_MOVE の reducer 内で makeMoveWithEffects が
  ターンチャージイベントを生成
- トラップ配置: BEGIN_PLAY_CARD → CONFIRM_PLAY_CARD で完結
- トラップ発動: makeMoveWithEffects 内の no_promote トラップ発動経路で完結

types.ts から 3 アクションを削除し、reducer.ts からも対応する case を
削除。trapTriggerEvent で参照している TrapTrigger 型は引き続き残す。

Issue #107 Step 5

* refactor: Step 5 — マナフライト/早指しバッジを専用フックに集約

card-shogi-game.tsx に散在していたマナ浮遊テキストと早指しバッジの
state / id ref / trigger / remove を、用途別のカスタムフックに集約する。
依存関係はローカルクローズドで責務が明確になる。

新規:
- use-mana-flight-layer.ts (26 行) — useManaFlightLayer フック
- use-fast-move-badge-layer.ts (25 行) — useFastMoveBadgeLayer フック

修正:
- card-shogi-game.tsx: 2 つの useState + 2 つの useRef + 4 つの useCallback
  をフック呼出 + 分割代入に置換 (約 25 行縮減)。trigger / remove の参照は
  既存名と同じになるよう分割代入時に rename して、ロジック側は無修正に

ロジック・振る舞い・呼出順は全て不変 (move-only)。

Issue #107 Step 5

* test: Step 5 — reducer のシナリオテスト 21 ケース追加

src/hooks/card-shogi/reducer.ts (Step 5 で抽出) の主要シナリオを
Vitest でカバー。リファクタ・将来の Action 追加に対する回帰防護。

カバー範囲 (21 ケース):
- 駒指し系: SELECT_SQUARE (自駒選択 / pendingCard ガード / isDrawing
  ガード) / DESELECT / RESIGN / SET_AI_THINKING /
  SHOW_PROMOTION_DIALOG / CANCEL_PROMOTION
- カード系 ドロー: DRAW_CARD (マナ十分 / マナ不足 / 手番外) / COMMIT_DRAW
- カード系 使用: BEGIN_PLAY_CARD (mana_up 成功 / マナ不足 / 手番外) /
  CONFIRM_PLAY_CARD (mana_up: マナ消費 + 手札→グレイブ + isPlayingCard) /
  CANCEL_PLAY_CARD / COMMIT_PLAY_CARD (currentPlayer 反転)
- UNDO: 履歴 < 2 ガード / カード操作介入時のガード
- RESET_TURN_TIMER: lastTurnStartedAt 設定

合計: 既存 effects.test.ts (70 ケース) と合わせて 91 ケース全パス。

Issue #107 Step 5
